requirements.txt
confluent_kafka
scrapy
protobuf
cassandra-driver